Add PersistableListAsObservable Remove ObservableList methods from PersistableList Let DisputeList and TradableList extend PersistableListAsObservable Fix param names
